+++ resources/assets/js/components/Estudiantes/CursosComponent.vue
@@ -21,9 +21,13 @@
          <template v-slot:item="props">
             <v-flex xs12 sm6 md4 lg3>
                <v-card>
-                  <v-card-title>
-                     <h4>{{ props.item.curso.nombre }}</h4>
-                  </v-card-title>
+                  <div :style="estilo()">
+                     <div class="img">
+                        <v-card-title text-sm-center>
+                           <h4 style="color: white;" class="headline">{{ props.item.curso.nombre }}</h4>
+                        </v-card-title>
+                     </div>
+                  </div>
                   <v-divider></v-divider>
                   <v-list dense>
                      <v-list-tile>
@@ -81,6 +85,26 @@ export default {
       },
       shownews() {
          if (this.total != "0") this.$router.push("/enoticias");
+      },
+      randomColor() {
+         let rc = "#";
+         for (let i = 0; i < 6; i++) {
+            rc += Math.floor(Math.random() * 9).toString(16);
+         }
+         return rc;
+      },
+      estilo() {
+         let color1 = this.randomColor();
+         let color2 = this.randomColor();
+         let salida = `background: #fc466b;
+                  background: -webkit-linear-gradient(
+                     to right,
+                     ${color1},
+                     ${color2}
+                  );
+                  background: linear-gradient(to right, ${color1}, ${color2});
+                  height: 120px;`;
+         return salida;
       }
    },
    mounted() {
@@ -116,5 +140,11 @@ export default {
    padding: 0px;
    margin: 0px;
    width: auto;
+}
+.img {
+   background-image: url("../../../img/textura.png");
+   height: 100%;
+   background-repeat: repeat;
+   padding: 0;
 }
 </style>
